requirements.txt
asgiref==3.7.1
Django==4.2.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
pillow==10.2.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.1
